python-src/requirement.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.2
setuptools==80.9.0
typing_extensions==4.15.0
wheel==0.45.1
